backend/requirements.txt
opencv-python~=4.10.0.84
mediapipe~=0.10.14
scikit-learn~=1.5.2
numpy>=1.21.0,<2.0.0
pyttsx3~=2.98
pillow~=10.4.0
flask
python-dotenv~=1.0.0
groq~=0.4.0
requests~=2.31.0
PyPDF2~=3.0.1
python-docx~=1.1.0
protobuf>=4.25.3,<5.0.0